Refactored so all test data is now in one file. Change was prompted by failing tests.
Made changes to failing tests.

Committing before starting work on actual ticket
