+++ app/src/androidTest/java/com/karhoo/karhootraveller/login/LoginRobot.kt
@@ -3,10 +3,10 @@ package com.karhoo.karhootraveller.login
 import com.karhoo.karhootraveller.R
 import com.karhoo.karhootraveller.common.BaseTestRobot
 import com.karhoo.karhootraveller.common.Launch
-import com.karhoo.karhootraveller.util.TestData
 import com.karhoo.uisdk.util.TestData.Companion.FILL_EMAIL
 import com.karhoo.uisdk.util.TestData.Companion.MEDIUM
 import com.karhoo.uisdk.util.TestData.Companion.SHOW_PASSWORD
+import com.karhoo.uisdk.util.TestData.Companion.USER_UPDATED
 
 fun login(func: LoginRobot.() -> Unit) = LoginRobot().apply {
     func()
@@ -63,7 +63,7 @@ class LoginRobot : BaseTestRobot() {
     fun fillValidEmailLogin() {
         fillEditText(
                 resId = R.id.emailInput,
-                text = TestData.USER_UPDATED.email
+                text = USER_UPDATED.email
                     )
     }
