+++ app/src/androidTest/java/com/karhoo/karhootraveller/menu/MenuTests.kt
@@ -11,7 +11,6 @@ import com.karhoo.karhootraveller.profile.user.userProfile
 import com.karhoo.uisdk.booking.booking
 import com.karhoo.uisdk.common.Launch
 import com.karhoo.uisdk.screen.booking.BookingActivity
-import com.karhoo.uisdk.util.TestData
 import com.karhoo.uisdk.util.TestData.Companion.LONG
 import org.junit.Rule
 import org.junit.Test
