+++ app/src/androidTest/java/com/karhoo/karhootraveller/registration/RegistrationTests.kt
@@ -7,10 +7,10 @@ import com.github.tomakehurst.wiremock.junit.WireMockRule
 import com.karhoo.karhootraveller.common.Launch
 import com.karhoo.karhootraveller.common.testrunner.TravellerTestConfig
 import com.karhoo.karhootraveller.presentation.register.RegistrationActivity
-import com.karhoo.uisdk.common.ServerRobot.Companion.EMAIL_ALREADY_IN_USE
-import com.karhoo.uisdk.common.ServerRobot.Companion.INVALID_PHONE_NUMBER
-import com.karhoo.uisdk.common.ServerRobot.Companion.REGISTRATION_FAILED
 import com.karhoo.uisdk.common.serverRobot
+import com.karhoo.uisdk.util.TestData.Companion.EMAIL_ALREADY_IN_USE
+import com.karhoo.uisdk.util.TestData.Companion.INVALID_PHONE_NUMBER
+import com.karhoo.uisdk.util.TestData.Companion.REGISTRATION_FAILED
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -73,6 +73,7 @@ class RegistrationTests : Launch {
         }
         registration(this) {
             enterRegistrationDetailsAlreadyUsedEmail()
+            waitFor(5000)
         } result {
             cannotRegisterUserErrorIsShown()
         }
