+++ app/src/androidTest/java/com/karhoo/karhootraveller/splash/SplashTests.kt
@@ -14,12 +14,12 @@ import com.karhoo.karhootraveller.presentation.splash.SplashActivity
 import com.karhoo.karhootraveller.profile.user.userProfile
 import com.karhoo.karhootraveller.registration.registration
 import com.karhoo.uisdk.booking.booking
-import com.karhoo.uisdk.common.ServerRobot.Companion.BRAINTREE_TOKEN
-import com.karhoo.uisdk.common.ServerRobot.Companion.REVERSE_GEO_SUCCESS
-import com.karhoo.uisdk.common.ServerRobot.Companion.USER_INFO
-import com.karhoo.uisdk.common.ServerRobot.Companion.USER_UPDATED_INFO
 import com.karhoo.uisdk.common.serverRobot
+import com.karhoo.uisdk.util.TestData.Companion.BRAINTREE_TOKEN
 import com.karhoo.uisdk.util.TestData.Companion.MEDIUM
+import com.karhoo.uisdk.util.TestData.Companion.REVERSE_GEO_SUCCESS
+import com.karhoo.uisdk.util.TestData.Companion.USER_INFO
+import com.karhoo.uisdk.util.TestData.Companion.USER_UPDATED_INFO
 import com.schibsted.spain.barista.rule.flaky.AllowFlaky
 import com.schibsted.spain.barista.rule.flaky.FlakyTestRule
 import org.junit.After
