+++ uisdk/src/androidTest/java/com/karhoo/uisdk/address/AddressTests.kt
@@ -7,15 +7,15 @@ import com.github.tomakehurst.wiremock.junit.WireMockRule
 import com.karhoo.uisdk.R
 import com.karhoo.uisdk.base.address.AddressType
 import com.karhoo.uisdk.common.Launch
-import com.karhoo.uisdk.common.ServerRobot.Companion.GENERAL_ERROR
-import com.karhoo.uisdk.common.ServerRobot.Companion.NO_ADDRESS_FOUND
-import com.karhoo.uisdk.common.ServerRobot.Companion.PLACE_SEARCH
-import com.karhoo.uisdk.common.ServerRobot.Companion.PLACE_SEARCH_AIRPORT
 import com.karhoo.uisdk.common.preferences
 import com.karhoo.uisdk.common.serverRobot
 import com.karhoo.uisdk.common.testrunner.UiSDKTestConfig
 import com.karhoo.uisdk.screen.address.AddressActivity
 import com.karhoo.uisdk.util.TestData
+import com.karhoo.uisdk.util.TestData.Companion.GENERAL_ERROR
+import com.karhoo.uisdk.util.TestData.Companion.NO_ADDRESS_FOUND
+import com.karhoo.uisdk.util.TestData.Companion.PLACE_SEARCH
+import com.karhoo.uisdk.util.TestData.Companion.PLACE_SEARCH_AIRPORT
 import com.karhoo.uisdk.util.TestData.Companion.SEARCH_AIRPORT_ADDRESS
 import com.karhoo.uisdk.util.TestData.Companion.SEARCH_GENERAL_ADDRESS
 import com.karhoo.uisdk.util.TestData.Companion.SEARCH_INCORRECT_ADDRESS
