+++ uisdk/src/androidTest/java/com/karhoo/uisdk/booking/BookingFlowTests.kt
@@ -8,17 +8,17 @@ import com.github.tomakehurst.wiremock.core.WireMockConfiguration.options
 import com.github.tomakehurst.wiremock.junit.WireMockRule
 import com.karhoo.uisdk.address.address
 import com.karhoo.uisdk.common.Launch
-import com.karhoo.uisdk.common.ServerRobot.Companion.BRAINTREE_TOKEN
-import com.karhoo.uisdk.common.ServerRobot.Companion.PLACE_DETAILS
-import com.karhoo.uisdk.common.ServerRobot.Companion.PLACE_SEARCH_RESULT
-import com.karhoo.uisdk.common.ServerRobot.Companion.REVERSE_GEO_SUCCESS
-import com.karhoo.uisdk.common.ServerRobot.Companion.REVERSE_GEO_SUCCESS_ALTERNATIVE
 import com.karhoo.uisdk.common.serverRobot
 import com.karhoo.uisdk.common.testrunner.UiSDKTestConfig
 import com.karhoo.uisdk.rides.rides
 import com.karhoo.uisdk.screen.booking.BookingActivity
 import com.karhoo.uisdk.util.TestData
+import com.karhoo.uisdk.util.TestData.Companion.BRAINTREE_TOKEN
 import com.karhoo.uisdk.util.TestData.Companion.MEDIUM
+import com.karhoo.uisdk.util.TestData.Companion.PLACE_DETAILS
+import com.karhoo.uisdk.util.TestData.Companion.PLACE_SEARCH_RESULT
+import com.karhoo.uisdk.util.TestData.Companion.REVERSE_GEO_SUCCESS
+import com.karhoo.uisdk.util.TestData.Companion.REVERSE_GEO_SUCCESS_ALTERNATIVE
 import com.karhoo.uisdk.util.TestData.Companion.SEARCH_ADDRESS
 import com.karhoo.uisdk.util.TestData.Companion.SEARCH_GENERAL_ADDRESS
 import com.schibsted.spain.barista.rule.flaky.AllowFlaky
