+++ uisdk/src/androidTest/java/com/karhoo/uisdk/rides/RidesTests.kt
@@ -6,32 +6,32 @@ import androidx.test.rule.ActivityTestRule
 import com.github.tomakehurst.wiremock.junit.WireMockRule
 import com.karhoo.uisdk.R
 import com.karhoo.uisdk.common.Launch
-import com.karhoo.uisdk.common.ServerRobot.Companion.DRIVER_TRACKING
-import com.karhoo.uisdk.common.ServerRobot.Companion.FARE_CANCELLED
-import com.karhoo.uisdk.common.ServerRobot.Companion.FARE_COMPLETE
-import com.karhoo.uisdk.common.ServerRobot.Companion.GENERAL_ERROR
-import com.karhoo.uisdk.common.ServerRobot.Companion.RIDE_SCREEN_CANCELLED_DRIVER
-import com.karhoo.uisdk.common.ServerRobot.Companion.RIDE_SCREEN_CANCELLED_KARHOO
-import com.karhoo.uisdk.common.ServerRobot.Companion.RIDE_SCREEN_CANCELLED_USER
-import com.karhoo.uisdk.common.ServerRobot.Companion.RIDE_SCREEN_COMPLETED
-import com.karhoo.uisdk.common.ServerRobot.Companion.RIDE_SCREEN_COMPLETED_AIRPORT_PICKUP
-import com.karhoo.uisdk.common.ServerRobot.Companion.RIDE_SCREEN_CONFIRMED
-import com.karhoo.uisdk.common.ServerRobot.Companion.RIDE_SCREEN_DER
-import com.karhoo.uisdk.common.ServerRobot.Companion.RIDE_SCREEN_DER_AIRPORT_PICKUP
-import com.karhoo.uisdk.common.ServerRobot.Companion.RIDE_SCREEN_INCOMPLETE
-import com.karhoo.uisdk.common.ServerRobot.Companion.RIDE_SCREEN_PREBOOKED
-import com.karhoo.uisdk.common.ServerRobot.Companion.RIDE_SCREEN_PREBOOKED_CANCELLED_BY_FLEET
-import com.karhoo.uisdk.common.ServerRobot.Companion.RIDE_SCREEN_PREBOOKED_CANCELLED_BY_USER
-import com.karhoo.uisdk.common.ServerRobot.Companion.TRIP_DER
-import com.karhoo.uisdk.common.ServerRobot.Companion.TRIP_HISTORY_EMPTY
-import com.karhoo.uisdk.common.ServerRobot.Companion.TRIP_STATUS_DER
 import com.karhoo.uisdk.common.serverRobot
 import com.karhoo.uisdk.common.testrunner.UiSDKTestConfig
 import com.karhoo.uisdk.ridedetail.rideDetail
 import com.karhoo.uisdk.screen.rides.RidesActivity
 import com.karhoo.uisdk.trip.trip
 import com.karhoo.uisdk.util.TestData
+import com.karhoo.uisdk.util.TestData.Companion.DRIVER_TRACKING
+import com.karhoo.uisdk.util.TestData.Companion.FARE_CANCELLED
+import com.karhoo.uisdk.util.TestData.Companion.FARE_COMPLETE
+import com.karhoo.uisdk.util.TestData.Companion.GENERAL_ERROR
 import com.karhoo.uisdk.util.TestData.Companion.LONG
+import com.karhoo.uisdk.util.TestData.Companion.RIDE_SCREEN_CANCELLED_DRIVER
+import com.karhoo.uisdk.util.TestData.Companion.RIDE_SCREEN_CANCELLED_KARHOO
+import com.karhoo.uisdk.util.TestData.Companion.RIDE_SCREEN_CANCELLED_USER
+import com.karhoo.uisdk.util.TestData.Companion.RIDE_SCREEN_COMPLETED
+import com.karhoo.uisdk.util.TestData.Companion.RIDE_SCREEN_COMPLETED_AIRPORT_PICKUP
+import com.karhoo.uisdk.util.TestData.Companion.RIDE_SCREEN_CONFIRMED
+import com.karhoo.uisdk.util.TestData.Companion.RIDE_SCREEN_DER
+import com.karhoo.uisdk.util.TestData.Companion.RIDE_SCREEN_DER_AIRPORT_PICKUP
+import com.karhoo.uisdk.util.TestData.Companion.RIDE_SCREEN_INCOMPLETE
+import com.karhoo.uisdk.util.TestData.Companion.RIDE_SCREEN_PREBOOKED
+import com.karhoo.uisdk.util.TestData.Companion.RIDE_SCREEN_PREBOOKED_CANCELLED_BY_FLEET
+import com.karhoo.uisdk.util.TestData.Companion.RIDE_SCREEN_PREBOOKED_CANCELLED_BY_USER
+import com.karhoo.uisdk.util.TestData.Companion.TRIP_DER
+import com.karhoo.uisdk.util.TestData.Companion.TRIP_HISTORY_EMPTY
+import com.karhoo.uisdk.util.TestData.Companion.TRIP_STATUS_DER
 import com.schibsted.spain.barista.rule.flaky.AllowFlaky
 import org.junit.After
 import org.junit.Rule
