+++ uisdk/src/androidTest/java/com/karhoo/uisdk/trip/TripLOCTest.kt
@@ -11,13 +11,15 @@ import com.github.tomakehurst.wiremock.junit.WireMockRule
 import com.karhoo.sdk.api.model.TripInfo
 import com.karhoo.uisdk.R
 import com.karhoo.uisdk.common.Launch
-import com.karhoo.uisdk.common.ServerRobot
 import com.karhoo.uisdk.common.networkServiceRobot
 import com.karhoo.uisdk.common.serverRobot
 import com.karhoo.uisdk.common.testrunner.UiSDKTestConfig
 import com.karhoo.uisdk.screen.trip.TripActivity
 import com.karhoo.uisdk.util.TestData
+import com.karhoo.uisdk.util.TestData.Companion.DRIVER_TRACKING
 import com.karhoo.uisdk.util.TestData.Companion.MEDIUM
+import com.karhoo.uisdk.util.TestData.Companion.TRIP_DER
+import com.karhoo.uisdk.util.TestData.Companion.TRIP_STATUS_DER
 import org.junit.After
 import org.junit.Rule
 import org.junit.Test
@@ -59,9 +61,9 @@ class TripLOCTest : Launch {
     @Test
     fun snackbarShowsToUserWhenWifiIsDisabled() {
         mockTripSuccessResponse(
-                status = ServerRobot.TRIP_STATUS_DER,
-                tracking = ServerRobot.DRIVER_TRACKING,
-                details = ServerRobot.TRIP_DER)
+                status = TRIP_STATUS_DER,
+                tracking = DRIVER_TRACKING,
+                details = TRIP_DER)
         trip(this) {
             networkServiceRobot {
                 disableNetwork(activityRule.activity.applicationContext)
